+++ src/generate_particle.cpp
@@ -51,7 +51,7 @@ int evaluateNumberParticles(GeomData &geomParams){
 void meshcube(GeomData &geomParams,
               SimulationData &simParams,
               vector<double> &pos,
-              vector<int> &type,
+              vector<double> &type,
               int &MP_count,
               int &FP_count){
 
@@ -129,7 +129,7 @@ void meshcube(GeomData &geomParams,
 void meshPostProcess(GeomData &geomParams,
                      SimulationData &simParams,
                      vector<double> &pos, 
-                     vector<int> &type,
+                     vector<double> &type,
                      int &GP_count){
 
     if(geomParams.post_process_do){
