+++ src/integration.h
@@ -19,8 +19,8 @@ void updateVariables(GeomData &geomParams,
                      vector<double> &gradW,
                      vector<double> &W,
                      vector<int> &neighbours,
-                     vector<int> &nb_neighbours,
-                     vector<int> type,
+                     vector<double> &nb_neighbours,
+                     vector<double> type,
                      vector<int> &track_particle);
 
 void checkTimeStep(GeomData &geomParams,    
@@ -30,4 +30,4 @@ void checkTimeStep(GeomData &geomParams,
                    vector<double> &u,
                    vector<double> &c,
                    vector<int> &neighbours,
-                   vector<int> &nb_neighbours);
+                   vector<double> &nb_neighbours);
